feat(bulk): nits
Refs; #1296
