+++ packages/api/src/command/medical/patient/bulk-doc-download-progress.ts
@@ -1,7 +1,11 @@
-import { DocumentDownloadStatus } from "../../../domain/medical/document-bulk-download";
+import {
+  DocumentDownloadStatus,
+  DocumentBulkDownloadProgress,
+} from "../../../domain/medical/document-bulk-download";
 import { Patient } from "../../../domain/medical/patient";
 import { PatientModel } from "../../../models/medical/patient";
 import { executeOnDBTx } from "../../../models/transaction-wrapper";
+import { BaseUpdateCmdWithCustomer } from "../base-update-command";
 import { getPatientOrFail } from "./get-patient";
 
 export type SetDocBulkDownloadProgress = {
@@ -19,15 +23,15 @@ export type SetDocBulkDownloadProgress = {
  * @returns a Promise that resolves to a Patient object.
  */
 export async function appendDocBulkDownloadProgress({
-  patient,
+  patient: { id, cxId },
   successful,
   errors,
   status,
   requestId,
 }: SetDocBulkDownloadProgress): Promise<Patient> {
   const patientFilter = {
-    id: patient.id,
-    cxId: patient.cxId,
+    id: id,
+    cxId: cxId,
   };
   return executeOnDBTx(PatientModel.prototype, async transaction => {
     const existingPatient = await getPatientOrFail({
@@ -67,3 +71,47 @@ export async function appendDocBulkDownloadProgress({
     return updatedPatient;
   });
 }
+
+export type BulkDownloadQueryInitCmd = BaseUpdateCmdWithCustomer & {
+  documentBulkDownloadProgress: Required<Pick<DocumentBulkDownloadProgress, "urlGeneration">>;
+  requestId: string;
+  totalDocuments?: number;
+};
+
+/**
+ * The function `storeBulkDownloadQueryInit` initalizes the `documentBulkDownloadProgress`field in a patient's data.
+ * @param {BulkDownloadQueryInitCmd} cmd - The `cmd` argument type to initialize the `documentBulkDownloadProgress` field
+ * @returns a Promise that resolves to a Patient object.
+ */
+export const storeBulkDownloadQueryInit = async (
+  cmd: BulkDownloadQueryInitCmd
+): Promise<Patient> => {
+  const { id, cxId, totalDocuments } = cmd;
+
+  return executeOnDBTx(PatientModel.prototype, async transaction => {
+    const patient = await getPatientOrFail({
+      id,
+      cxId,
+      lock: true,
+      transaction,
+    });
+
+    const update = {
+      documentBulkDownloadProgress: {
+        ...cmd.documentBulkDownloadProgress,
+        total: totalDocuments,
+      },
+      requestId: cmd.requestId,
+    };
+
+    return patient.update(
+      {
+        data: {
+          ...patient.data,
+          ...update,
+        },
+      },
+      { transaction }
+    );
+  });
+};
