+++ packages/api/src/command/medical/patient/query-init.ts
@@ -1,5 +1,4 @@
 import { DocumentQueryProgress } from "../../../domain/medical/document-query";
-import { DocumentBulkDownloadProgress } from "../../../domain/medical/document-bulk-download";
 import { Patient } from "../../../domain/medical/patient";
 import { QueryProgress } from "../../../domain/medical/query-status";
 import { PatientModel } from "../../../models/medical/patient";
@@ -42,50 +41,6 @@ export const storeQueryInit = async (cmd: QueryInitCmd): Promise<Patient> => {
           consolidatedQuery: cmd.consolidatedQuery,
           cxConsolidatedRequestMetadata: cmd.cxConsolidatedRequestMetadata,
         };
-
-    return patient.update(
-      {
-        data: {
-          ...patient.data,
-          ...update,
-        },
-      },
-      { transaction }
-    );
-  });
-};
-
-export type BulkDownloadQueryInitCmd = BaseUpdateCmdWithCustomer & {
-  documentBulkDownloadProgress: Required<Pick<DocumentBulkDownloadProgress, "download">>;
-  requestId: string;
-  totalDocuments?: number;
-};
-
-/**
- * The function `storeBulkDownloadQueryInit` initalizes the `documentBulkDownloadProgress`field in a patient's data.
- * @param {BulkDownloadQueryInitCmd} cmd - The `cmd` argument type to initialize the `documentBulkDownloadProgress` field
- * @returns a Promise that resolves to a Patient object.
- */
-export const storeBulkDownloadQueryInit = async (
-  cmd: BulkDownloadQueryInitCmd
-): Promise<Patient> => {
-  const { id, cxId, totalDocuments } = cmd;
-
-  return executeOnDBTx(PatientModel.prototype, async transaction => {
-    const patient = await getPatientOrFail({
-      id,
-      cxId,
-      lock: true,
-      transaction,
-    });
-
-    const update = {
-      documentBulkDownloadProgress: {
-        ...cmd.documentBulkDownloadProgress,
-        total: totalDocuments,
-      },
-      requestId: cmd.requestId,
-    };
 
     return patient.update(
       {
